refactor(openomni): fold tool/shared input/result helpers into define.ts
The shared/ dir held builtin authoring helpers with define.ts as the
natural home (every builtin already imports defineTool). Tests merged
into define.test.ts; no shared name survives.
